Add files via upload
The code begins by importing the random and time modules.

The obter_leitura_temperatura() function is defined. This function simulates the reading of the temperature by generating a random value between 60 and 100 degrees Celsius using the random.uniform() function. It then returns the generated temperature value.

The monitorar_temperatura() function is defined. This function contains a while loop that continuously monitors the temperature.

Inside the loop, the obter_leitura_temperatura() function is called to obtain a temperature reading. The obtained temperature is then printed to the console using the print() function.

After printing the temperature, an if statement checks if the temperature is greater than 90 degrees. If it is, an alert message "ALERTA: Temperatura alta detectada!" (which means "ALERT: High temperature detected!" in Portuguese) is printed to the console.

Following the alert message, you can add your own logic to notify the maintenance team or take any other necessary actions.

After the if statement, the code pauses execution for 1 second using the time.sleep() function. This delay ensures that there is a 1-second interval between temperature readings.

The monitorar_temperatura() function is called at the end, which initiates the temperature monitoring process. Since the while loop is set to run indefinitely (while True:), the code will continue monitoring the temperature until it is interrupted externally.

Overall, this code continuously generates random temperature readings, prints them to the console, and raises an alert if the temperature exceeds 90 degrees Celsius. The time.sleep() function introduces a delay between each temperature reading to control the frequency of monitoring.





Still working on this :)